app-ai/requirements.txt
python-dotenv==1.1.1
openai==1.97.0